pyproject.toml
[project]
name = "qualifybot"
version = "0.1.0"
description = "IT Support Voice Assistant - Automated IT helpdesk via phone calls"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "redis>=5.2.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "openai>=1.54.0",
    "elevenlabs>=1.0.0",
    "twilio>=9.3.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "sentry-sdk>=2.19.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "langchain-openai>=0.0.2",
    "langgraph-checkpoint-redis>=0.3.0",
    "pydub>=0.25.1",
    "instructor>=1.0.0",
    "chromadb>=0.4.0",
    "langchain-chroma>=0.1.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "jira>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.8.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "httpx>=0.27.0",  # for testing
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.8.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
# Dev dependencies moved to dependency-groups (new format)

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "DTZ", "T10", "SIM", "ARG", "PIE", "PT", "RET", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

